Update call to initialize cohort registry to ensure cohort registry store is alway passed.
This is in response to the following note from Ruxandra Rosu

Hi Mandy,

I am trying to initialize the cohort for the server running OMAS’s, but I am having some issues. For the IGCEventMapper and Atlas servers it seems to work.

I am using the below calls, but I get OMRSErrorCode.NULL_REGISTRY_STORE exception when initializing the cohort because the registryStore is null (OMRSCohortRegistry – line 104). This is because localRepository is null- line 129 in OMRSCohortManager.

Thanks,
Ruxandra

Signed-off-by: Mandy Chessell <mandy_chessell@uk.ibm.com>