Save local repository's metadata collection Id
When the enterprise connector is requested to update metadata,
it needs to identify which store to redirect the request to.
It does this by comparing the home collection Id in the
instance with the metadata collection id assigned to each repository in the cohort.
In the logic today, the code that is supposed to set up the variable containing the
local repository's metadata collection id is actually setting its own metadata collection
id in the super class rather than the local variable.

This means that the enterprise connector is not able to find the repository that corresponds
to the local repository.

This is covered by Jira https://jira.odpi.org/browse/EGERIA-28

Signed-off-by: Mandy Chessell <mandy_chessell@uk.ibm.com>