Ensure audit log is passed to the cohort topic connector - and its initialization is moved to start() so that the audit log is available.
This commit also moves the sending of events over Kafka to a separate thread so start up
does not wait for completion of registration before returning.  Error messages are output
approximately every 10 minutes.

Also there is part of the implementation of the Connected Asset OMAS as I was partway through this
change.  This includes changes to the OCF to support API schemas

Signed-off-by: Mandy Chessell <mandy_chessell@uk.ibm.com>