GE OMAS - Initial Commit
GE OMAS provides API to query from governed assets available in the repository.
In this moment, only the Relational Column, Relational Table and Glossary terms are processed (maybe the REST endpoint should be renamed and the Context object should be more generic).
On the Output topic are send events for Classified Entity and New Relationship (Semantic Assignment).

Signed-off-by: Daniela Otelea <daniela-valentina.otelea@ing.com>