Changes to OMRSRepositoryConnector API.
Added ability for a repository connector to return FunctionNotReportedException on instance
metadata maintenance calls (such as addEntity).

Change affected OMASs to convert this exception into RepositoryErrorException.

Deprecated findXXXByProperty methods and replaces them by new version that allow the definition
of search criteria that must match exactly or be a regEx style fuzzy search.

Added new method called findXXXByExactPropertyValue for finding exact matches.

The changes to the find methods where to clarify whether special characters need to be
treated as part of the search string or part of the property value.

Signed-off-by: Mandy Chessell <mandy_chessell@uk.ibm.com>