Add support for capturing and logging unexpected exceptions
The new method and message definition supports the capture of an unexpected exception during the server-side execution of a REST API call.  The exception information is copied into the response object and an EXCEPTION audit log event is pushed to the AuditLog.  This event includes the stack trace of the error.
However this stack trace is not sent to thhe client.

Signed-off-by: Mandy Chessell <mandy_chessell@uk.ibm.com>