Add Dependency on later version of Sleepycat
The current version of janusgraph-berkeleyje (a dependency of
JanusGraph) relies on an outdated version of com.sleepycat.je that while
available from Oracle's Maven repository[1], it is no longer replicated
to Maven Central.

Thus when building Egeria without a cached version of sleepycat (or
pulling from a Maven repository without the cached version), some users
will continue to experience failures when building Egeria. Specifing a
greater version of sleepycat appears to bypass this failure and allow
the build to succeed.

Resolves: #1116

[1] https://download.oracle.com/maven/com/sleepycat/je/7.4.5/je-7.4.5.pom
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>