Run Nexus IQ Scans on Merges
Adds a job to run Nexus IQ scans each time code is merged and on a
weekly basis, so that users can investigate the reports and take action.

There is a Nexus IQ task for Azure Pipelines which runs the nexus-iq cli
jar, but it doesn't not provide enough configuration options currently
to run correctly. It fails with an E2BIG error because it tries to pass
the filepaths of all jars on the command line, which ends up being more
characters than are allowed.

The version 2.8.1 is used as the IQ server version isn't compatible with
the newest IQ scanner plugin, but can be bumped once the server is
updated.

Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>