Switch Artifact Names to use BuildNumber
It's still unclear what caused this issue: a build failed with the error
"Artifact .. already exists for build $(Build.BuildId)".

The documentation for AZP is not clear how BuildId is set or calculated,
but we can have more confidence that the artifact names will be unique
when using BuildName which might stop this from occuring in the future.

Resolves #1703
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>