Store the Settings file locally in the Repository
Instead of using Azure Secret Files, which can't be viewed or edited
without overwriting the entire file, credentials for servers are
stored in the AZP variable library and pulled into a global settings
file stored locally in the repostitory.

This makes it clearer what the contents of the settings.xml file is for
Egeria and allows for extra configuration to be added in the future
without the need to overwrite the entire file.

Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>