Add Properties for Node and NPM Download URLs
The properties are set to the same defaults as the upstream plugin for
the node and npm download roots.

By creating properties for the frontend-maven-plugin's nodeDownloadRoot
and npmDownloadRoot configuration, the CI will be able to override the
values and point at a proxied location while not impacting anyone
building the project locally.

Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>