Search Indexing Connector
This is a draft PR. The implementation is not finished. I would like to check if the current implementation follows the pattern used for the OMRS connectors.
The purpose of this connector is to provide an efficient way to find entities by type from the cohort.
The configuration of the connector should provide the connection to Elastic search and the OM types used in for Enterprise OMRS topic events (the processing of these events should be based on the types: for PoC we choose Asset, Schema Elements, Glossary Terms).
The first part of the search indexing connector is part of the OMRS layer, the actual implementation of it is adapters - repository services.

This is an example of request body for connector config: {{server-url}}/open-metadata/admin-services/users/{{user-id}}/servers/{{server-id}}/search-indexing/configuration
{
    "class":"SearchIndexingConfig",
    "description":"Example Search Indexing Solution",
    "searchIndexingConnection":{
        "class":"Connection",
        "displayName":"Search Indexing Connection",
        "description":"Used for storing and fetching metadata in the Open Metadata format",
        "connectorType":{
            "class":"ConnectorType",
            "connectorProviderClassName":"org.odpi.openmetadata.adapters.repositoryservices.searchindexing.repositoryconnector.SearchIndexingRepositoryConnectorProvider"
        },
        "endpoint":{
        	"class": "Endpoint",
        	"address": "{{search-host}}:{{search-port}}"
        },
        "configurationProperties":{
            "storageBackend":"cassandra",
            "indexSearchBackend":"elastic-search"
        }
    }
}

Signed-off-by: Daniela Otelea <daniela-valentina.otelea@ing.com>