Update egeria-dojo-day-1-4-participating-in-the-community.md
As reported by user on Slack looks like the "Code of conduct" link pointing to the .md file does show the page (suspect because of the different naming convention and internals of how it is translated to html..). Updating the link to the actual github location like it is already in the other dojo pages.

Signed-off-by: Ljupcho Palashevski <ljupcho.palashevski@ing.com>