Change parameter in function calls
The notebook called
queryKnownServers(platformName, platformServicesURLRoot)
queryActiveServers(platformName, platformServicesURLRoot)

The functions in common expects the url+portnumber which is in
platformURLroot.
Then function will internally generate platformServicesURLRoot.

Renamed platformURLroot to serverPlatformURL to align with the
name used internally in the function.
Changed function call parameter to serverPlatformURL in
place of platformServicesURLRoot

Signed-off-by: Theo Thomas <tcn.thomas@googlemail.com>